refactor: solving SonarCloud codeSmells and severity
