+++ src/main/java/com/softserveinc/dokazovi/controller/AuthController.java
@@ -3,10 +3,10 @@ package com.softserveinc.dokazovi.controller;
 import com.softserveinc.dokazovi.entity.UserEntity;
 import com.softserveinc.dokazovi.entity.enumerations.UserStatus;
 import com.softserveinc.dokazovi.exception.BadRequestException;
-import com.softserveinc.dokazovi.payload.ApiResponse;
-import com.softserveinc.dokazovi.payload.AuthResponse;
-import com.softserveinc.dokazovi.payload.LoginRequest;
-import com.softserveinc.dokazovi.payload.SignUpRequest;
+import com.softserveinc.dokazovi.entity.payload.ApiResponse;
+import com.softserveinc.dokazovi.entity.payload.AuthResponse;
+import com.softserveinc.dokazovi.entity.payload.LoginRequest;
+import com.softserveinc.dokazovi.entity.payload.SignUpRequest;
 import com.softserveinc.dokazovi.security.TokenProvider;
 import com.softserveinc.dokazovi.service.ProviderService;
 import com.softserveinc.dokazovi.service.UserService;
