+++ src/main/java/com/softserveinc/dokazovi/security/TokenAuthenticationFilter.java
@@ -24,7 +24,8 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
     @Autowired
     private CustomUserDetailsService customUserDetailsService;
 
-    private static final Logger logger = LoggerFactory.getLogger(TokenAuthenticationFilter.class);
+    private static final Logger tokenAuthenticationFilterLogger = LoggerFactory
+            .getLogger(TokenAuthenticationFilter.class);
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
@@ -41,7 +42,7 @@ public class TokenAuthenticationFilter extends OncePerRequestFilter {
                 SecurityContextHolder.getContext().setAuthentication(authentication);
             }
         } catch (Exception ex) {
-            logger.error("Could not set user authentication in security context", ex);
+            tokenAuthenticationFilterLogger.error("Could not set user authentication in security context", ex);
         }
         filterChain.doFilter(request, response);
     }
