+++ src/main/java/com/softserveinc/dokazovi/security/oauth2/user/OAuth2UserInfo.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public abstract class OAuth2UserInfo {
     protected Map<String, Object> attributes;
 
-    public OAuth2UserInfo(Map<String, Object> attributes) {
+    protected OAuth2UserInfo(Map<String, Object> attributes) {
         this.attributes = attributes;
     }
