+++ src/main/java/com/softserveinc/dokazovi/security/oauth2/user/OAuth2UserInfoFactory.java
@@ -9,6 +9,9 @@ import java.util.Map;
 
 public class OAuth2UserInfoFactory {
 
+    private OAuth2UserInfoFactory() {
+    }
+
     public static OAuth2UserInfo getOAuth2UserInfo(String registrationId, Map<String, Object> attributes) {
         if (registrationId.equalsIgnoreCase(AuthProvider.google.toString())) {
             return new GoogleOAuth2UserInfo(attributes);
